fix(parse/css): let `@utility` parse declarations and rules
