+++ crates/biome_css_parser/tests/css_test_suite/ok/tailwind/utility/with-sub-block.css
@@ -1,0 +1,7 @@
+@utility border-overlay-* {
+	position: relative;
+
+	&::after {
+		border-width: 1px;
+	}
+}
